+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This changelog track changes to the qoqo project starting at version v0.5.0
 
 ### Added in 1.6.0-alpha.0
 
-* Added first bosonic operations traits and derives in `roqoqo` and `roqoqo-derive`.
+* Added first bosonic operations traits in `roqoqo` and `qoqo`.
 
 ## 1.5.0